+++ CHANGES.md
@@ -1,4 +1,15 @@
-### 1.11.0
+## 1.12.0
+
+An enhanced API supporting the use of
+`dynamodb.write_versioned.versioned_transact_write_items`, which is a
+great way to write business logic against DynamoDB.
+
+### 1.11.1
+
+Change all Resources and Clients to use a thread-local Session to
+avoid race conditions.
+
+## 1.11.0
 
 - Update `find_index` and `require_index` to allow for an index with no range key.